refactor: other_events.ts を common/member_list.ts にリネームして移動
